requirements.txt
# local package
-e .

# external requirements
torch   
torchvision   
pytorch-lightning   
torchmetrics
optuna   
optuna-dashboard   
tensorboard   
hydra-core   
hydra-joblib-launcher    
hydra-optuna-sweeper    
tqdm   
matplotlib  
seaborn  
snakeviz  
ipykernel  
pytest